// formularioReactTiendasEl/ComplaintsBook/react/typings/complaints-book.d.ts
export interface ComplaintsBookForm {
    time: string;
    correlativeNumber: string;
    storeName: string;
    storeCode: string;
    storeAddress: string;
    rucStoreNumber: string;
    name: string;
    lastName: string;
    documentType: string;
    documentNumber: any;
    departamento: string;
    provincia: string;
    distrito: string;
    address: string;
    phone: any;
    email: string;
    representative: string;
    requestDetails: string;
    requestNumber: any;
    requestPrice: any;
    claimDetails: string;
    claimDescription: string;
    clientOrder: string;
    providerActions: string;
}